Tidy code

+++ src/lib.rs
@@ -7,4 +7,4 @@ pub use interface::*;
 pub use settings::*;
 pub const CPU_CYCLES_PER_SCANLINE: f32 = 113.67;
 pub const CPU_CYCLES_PER_VBLANK: i64 = 2260;
-pub const CPU_CYCLES_PER_OAM: f32 = 513.0;
+pub const CPU_CYCLES_PER_OAM: u32 = 513;
